requirements.txt
numpy>=1.20.0
numba>=0.53.0
pillow>=8.0.0
scipy>=1.6.0
matplotlib>=3.3.0
cupy-cuda11>=10.0.0
torch>=1.9.0
opencv-python>=4.5.0
scikit-image>=0.18.0
seaborn>=0.11.0
